+++ CHANGELOG.md
@@ -60,6 +60,7 @@ Compatibility:
 * Fix `Range#step` with no block and non-`Numeric` values (#2824, @eregon).
 * Fix execution order of `END` blocks and `at_exit` callbacks (#2818, @andrykonchin).
 * Fix `String#casecmp?` for empty strings of different encodings (#2826, @eregon).
+* Implement `Enumerable#compact` and `Enumerator::Lazy#compact` (#2733, @andrykonchin).
 
 Performance: